requirements.txt
# MealTrace Digital — Backend Dependencies
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Firebase
firebase-admin==6.6.0

# Environment & Config
python-dotenv==1.0.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# Authentication
PyJWT==2.10.1
cryptography==44.0.0

# QR Code Generation
qrcode[pil]==8.0
Pillow==11.1.0

# Payment Gateway
razorpay==2.0.1
setuptools>=70.0.0

# Excel Export (Reports)
openpyxl==3.1.5

# HTTP Client (for async operations)
httpx==0.28.1

# Date/Time Utilities
python-dateutil==2.9.0